backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
supabase==2.0.2
redis==5.0.1

# AI and ML - Only using Claude 4
anthropic==0.7.8  # Claude 4
httpx==0.24.1     # For Vapi and Fetch AI API calls

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
cryptography==41.0.7 